Completed refactoring check unit tests
