+++ System/Directory.Build.props
@@ -1,19 +1,19 @@
 <Project>
-  <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory)..\, Directory.Build.props))\Directory.Build.props" />
-  <PropertyGroup>
-    <VersionPrefix>2.0.0</VersionPrefix>
-    <Title>Wangkanai System</Title>
-    <PackageTags>aspnetcore;sytem;runtime;</PackageTags>
-    <Description>Wangkanai .NET Core System library to serve as the foundation for future development.</Description>
-    <RootNamespace>Wangkanai</RootNamespace>
-    <PackageProjectUrl>https://github.com/wangkanai/wangkanai/tree/main/System</PackageProjectUrl>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(MSBuildProjectName)' == 'Wangkanai.System'">
-    <PackageIcon>wangkanai-logo.png</PackageIcon>
-    <PackageReadmeFile>README.md</PackageReadmeFile>
-  </PropertyGroup>
-  <ItemGroup Condition="'$(MSBuildProjectName)' == 'Wangkanai.System'">
-    <None Include="..\..\assets\wangkanai-logo.png" Pack="true" PackagePath="\" />
-    <None Include="..\README.md" Pack="true" PackagePath="\" />
-  </ItemGroup>
+    <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory)..\, Directory.Build.props))\Directory.Build.props"/>
+    <PropertyGroup>
+        <VersionPrefix>2.0.0</VersionPrefix>
+        <Title>Wangkanai System</Title>
+        <PackageTags>aspnetcore;sytem;runtime;</PackageTags>
+        <Description>Wangkanai .NET Core System library to serve as the foundation for future development.</Description>
+        <RootNamespace>Wangkanai</RootNamespace>
+        <PackageProjectUrl>https://github.com/wangkanai/wangkanai/tree/main/System</PackageProjectUrl>
+    </PropertyGroup>
+    <PropertyGroup Condition="'$(MSBuildProjectName)' == 'Wangkanai.System'">
+        <PackageIcon>wangkanai-logo.png</PackageIcon>
+        <PackageReadmeFile>README.md</PackageReadmeFile>
+    </PropertyGroup>
+    <ItemGroup Condition="'$(MSBuildProjectName)' == 'Wangkanai.System'">
+        <None Include="..\..\assets\wangkanai-logo.png" Pack="true" PackagePath="\"/>
+        <None Include="..\README.md" Pack="true" PackagePath="\"/>
+    </ItemGroup>
 </Project>
