+++ System/tests/Checks/CheckNumericTests.cs
@@ -1,0 +1,35 @@
+﻿// Copyright (c) 2014-2022 Sarin Na Wangkanai, All Rights Reserved.Apache License, Version 2.0
+
+#nullable enable
+using Xunit;
+
+namespace Wangkanai.Checks;
+
+public class CheckNumericTests
+{
+	[Fact]
+	public void IntegralIsNull()
+	{
+		byte?  byte1  = null;
+		short? short2 = null;
+		int?   int4   = null;
+		long?  long8  = null;
+
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(byte1));
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(short2));
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(int4));
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(long8));
+	}
+
+	[Fact]
+	public void FloatingIsNull()
+	{
+		float?   float16   = null;
+		double?  double32  = null;
+		decimal? decimal32 = null;
+
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(float16));
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(double32));
+		Assert.Throws<ArgumentNullException>(() => Check.ThrowIfNull(decimal32));
+	}
+}
