+++ System/tests/Cryptography/Adler32Tests.cs
@@ -15,7 +15,7 @@ public class Adler32Tests
 		Assert.Throws<ArgumentLessThanException>(() => Adler32.ComputeChecksum(-1, new byte[0], -1, 0));
 		Assert.Throws<ArgumentLessThanException>(() => Adler32.ComputeChecksum(-1, new byte[0], 0, -1));
 	}
-	
+
 	[Fact]
 	public void Checksum_Normal()
 	{
@@ -28,10 +28,10 @@ public class Adler32Tests
 	{
 		Assert.Throws<ArgumentNullException>(() => Adler32.ComputeChecksum(0, null!));
 	}
-	
+
 	[Fact]
-    public void Checksum_Empty()
-    {
-        //Assert.Throws<ArgumentZeroException>(() => Adler32.ComputeChecksum(0, new byte[0], 0, 0));
-    }
+	public void Checksum_Empty()
+	{
+		//Assert.Throws<ArgumentZeroException>(() => Adler32.ComputeChecksum(0, new byte[0], 0, 0));
+	}
 }
